api/requirements.txt
django
djangorestframework
django-cors-headers
gunicorn
requests
bcrypt